+++ demo/module3/src/main/java/com/sankuai/erp/component/appinitmodule3/init/Module3TwoInit.kt
@@ -9,7 +9,7 @@ import com.sankuai.erp.component.appinit.common.AppInitLogger
  * 创建时间:2018/11/28
  * 描述:
  */
-@AppInit(priority = 60, description = "模块32的描述", aheadOf = "com.sankuai.erp.component:appinit-test-module1:Module1OneInit")
+@AppInit(priority = 60, description = "模块32的描述", aheadOf = "cn.bingoogolapple:appinit-test-module1:Module1OneInit")
 class Module3TwoInit : SimpleAppInit() {
     override fun onCreate() {
         AppInitLogger.demo("onCreate $TAG")